working datetime64 with some conversions
