+++ dbms/src/DataTypes/DataTypeDateTime.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <Core/DateTime64.h>
 #include <DataTypes/DataTypeNumberBase.h>
 
 
@@ -72,27 +73,6 @@ protected:
 
 struct DataTypeDateTime : DataTypeDateTimeBase<UInt32> {
     using DataTypeDateTimeBase::DataTypeDateTimeBase;
-};
-
-// this is a separate class to avoid accidental conversions that
-// might occur between time_t and the type storing the datetime64
-// time_t might have a different definition on different libcs
-struct DateTime64 {
-    using Type = Int64;
-    struct Components {
-        time_t datetime = 0;
-        UInt32 nanos = 0;
-    };
-
-    Components split() const;
-    explicit DateTime64(Components c);
-    explicit DateTime64(Type tt) : t{tt} {}
-    explicit operator bool() const {
-        return t != 0;
-    }
-    Type get() const { return t; }
-private:
-    Type t;
 };
 
 struct DataTypeDateTime64 : DataTypeDateTimeBase<DateTime64::Type> {
