+++ dbms/tests/queries/0_stateless/00921_datetime64.reference
@@ -1,3 +1,2 @@
-2	1970-01-01 03:00:00.000000001	1	0
-2	1970-01-01 03:00:00.000050000	3	3
-2	2019-05-03 02:25:25.123456789	5	3
+2019-05-03 10:25:25.123456789	2019-05-03	2019-05-03 00:00:00	2019-04-01	1970-01-02 11:25:25	2019-05-03 11:25:00
+2019-05-03 10:25:25.123456789	2019-05-03	2019-05-03 00:00:00	2019-04-01	1970-01-02 11:25:25	2019-05-03 11:25:00
